💄 Add new color palette
The background-color of the NavBar component is now transparent, which
improves the visual appearance of the component. The color of the links
is now set to currentColor, which allows the color to inherit from its
parent element. The color and background-color variables in the globals.scss
file are now changed to $text-100 and $bg-100, respectively, which improves
the readability of the text and the overall appearance of the application.
New color variables are added to the _colors.scss file to improve the
consistency of the color scheme.