🚧 Set styles for Login page and Register page
The primary and secondary buttons no longer have a hover effect. The Login and SignUp modules now use a container mixin to apply a consistent style. The SignUpForm component now has a default option for the gender and document type select inputs. The login.tsx and signup.tsx pages now use a title style for the h1 element. A container mixin has been